+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy/A32NX_FMCMainDisplay.ts
@@ -54,7 +54,7 @@ import { FmgcFlightPhase } from '@shared/flightphase';
 import { CompanyRoute } from '@simbridge/index';
 import { Keypad } from './A320_Neo_CDU_Keypad';
 import { FmsClient } from '@atsu/fmsclient';
-import { AtsuStatusCodes, CruiseWindRequest, WindRequestMessage } from '@datalink/common';
+import { AtsuStatusCodes } from '@datalink/common';
 import { A320_Neo_CDU_MainDisplay } from './A320_Neo_CDU_MainDisplay';
 import { FmsDisplayInterface } from '@fmgc/flightplanning/interface/FmsDisplayInterface';
 import { FmsError, FmsErrorType } from '@fmgc/FmsError';
@@ -66,7 +66,6 @@ import {
   SimVarValueType,
   ClockEvents,
   ConsumerSubject,
-  Vec2Math,
   ConsumerValue,
   Accessible,
   Subject,
@@ -78,7 +77,6 @@ import { FmsFormatters } from './FmsFormatters';
 import { NavigationDatabase, NavigationDatabaseBackend } from '@fmgc/NavigationDatabase';
 import { FlightPhaseManager } from '@fmgc/flightphase';
 import { FlightPlanService } from '@fmgc/flightplanning/FlightPlanService';
-import { HistoryWind } from '@fmgc/wind/HistoryWind';
 import { DefaultPerformanceData } from '@fmgc/flightplanning/plans/performance/FlightPlanPerformanceData';
 import { A320FlightPlanPerformanceData } from '@fmgc/flightplanning/plans/performance/A320FlightPlanPerformanceData';
 import { NavigationDatabaseService } from '@fmgc/flightplanning/NavigationDatabaseService';
@@ -96,9 +94,6 @@ import { CDUFlightPlanPage } from '../legacy_pages/A320_Neo_CDU_FlightPlanPage';
 import { FuelPredComputations } from '@fmgc/flightplanning/fuel/FuelPredComputations';
 import { MsfsFlightPlanSync } from '@fmgc/flightplanning/MsfsFlightPlanSync';
 import { PendingWindUplinkParser } from '@fmgc/flightplanning/plans/PendingWindUplinkParser';
-import { isLeg, FlightPlanLeg } from '@fmgc/flightplanning/legs/FlightPlanLeg';
-import { ProfilePhase } from '@fmgc/guidance/vnav/profile/NavGeometryProfile';
-import { SegmentClass } from '@fmgc/flightplanning/segments/SegmentClass';
 import { bearingTo } from 'msfs-geo';
 import { WindUtils } from '@fmgc/guidance/vnav/wind/WindUtils';
 import { EngineOutControlEvents, EngineOutEvents } from '@fmgc/events/EngineOutEvents';
@@ -108,6 +103,7 @@ import { FlightPlan } from '@fmgc/flightplanning/plans/FlightPlan';
 import { A32NXFcuBusEvents } from '@shared/publishers/A32NXFcuBusPublisher';
 import { A32NXFgBusEvents } from '@shared/publishers/A32NXFGBusPublisher';
 
+import { formatWindRequest } from '@fmgc/flightplanning/uplink/WindUplinkUtilts';
 export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInterface, Fmgc {
   private static DEBUG_INSTANCE: FMCMainDisplay;
 
@@ -140,7 +136,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
   public readonly currFlightPlanService = new FlightPlanService(
     this.bus,
     new A320FlightPlanPerformanceData(),
-    FpmConfigs.A320_HONEYWELL_H3,
+    FpmConfigs.A320_HONEYWELL_H4,
     true,
   );
   private readonly observableFlightPlanManager = new ObservableFlightPlanManager(this.bus, this.currFlightPlanService);
@@ -171,7 +167,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
 
   /** Declaration of every variable used (NOT initialization) */
   private readonly maximumAllowedCruiseFlightLevel = 390;
-  private readonly maximumRecommendedCruiseFlightLevel = 398;
+  public static readonly maximumRecommendedCruiseFlightLevel = 398;
   public coRoute = { routeNumber: undefined, routes: undefined };
 
   private readonly fuelComputationsCache: Map<FlightPlanIndex, FuelPredComputations> = new Map();
@@ -327,7 +323,6 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
   public efisInterfaces?: Record<EfisSide, EfisInterface>;
   public guidanceController?: GuidanceController;
   public navigation?: Navigation;
-  private historyWinds?: HistoryWind;
 
   public casToMachManualCrossoverCurve;
   public machToCasManualCrossoverCurve;
@@ -532,7 +527,6 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
       this.efisInterfaces.R,
       a320EfisRangeSettings,
     );
-    this.historyWinds = new HistoryWind(this.bus);
 
     initComponents(this.bus, this.navigation, this.guidanceController, this.flightPlanService);
 
@@ -2369,6 +2363,9 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
       case FmsErrorType.AwyWptMismatch:
         this.setScratchpadMessage(NXSystemMessages.awyWptMismatch);
         break;
+      case FmsErrorType.NotAllowed:
+        this.setScratchpadMessage(NXSystemMessages.notAllowed);
+        break;
     }
   }
 
@@ -3698,14 +3695,20 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     return true;
   }
 
-  public setPerfApprWind(s: string, forPlan: FlightPlanIndex): boolean {
+  public async setPerfApprWind(s: string, forPlan: FlightPlanIndex): Promise<boolean> {
+    const plan = this.flightPlanService.has(forPlan) ? this.flightPlanService.get(forPlan) : null;
+
+    if (!plan || plan.destinationAirport === undefined) {
+      this.setScratchpadMessage(NXSystemMessages.notAllowed);
+      return false;
+    }
+
     if (s === Keypad.clrValue) {
-      this.flightPlanService.setPerformanceData('approachWindDirection', null, forPlan);
-      this.flightPlanService.setPerformanceData('approachWindMagnitude', null, forPlan);
-      this.flightPlanService.setPerformanceData('isApproachWindPilotEntered', false, forPlan);
-      this.flightPlanService.setDescentWindEntry(0, null, forPlan);
-      this.updateTowerHeadwind(null, null, null);
-      return true;
+      const success = await this.flightPlanService.deleteApproachWind(forPlan);
+      if (success && forPlan === FlightPlanIndex.Active) {
+        this.updateTowerHeadwind(null, null, null);
+      }
+      return success;
     }
 
     // both must be entered
@@ -3718,24 +3721,12 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
       this.setScratchpadMessage(NXSystemMessages.entryOutOfRange);
       return false;
     }
-    this.flightPlanService.setPerformanceData('approachWindDirection', dir % 360, forPlan); // 360 is displayed as 0
-    this.flightPlanService.setPerformanceData('approachWindMagnitude', mag, forPlan);
-    this.flightPlanService.setPerformanceData('isApproachWindPilotEntered', true, forPlan);
 
-    const plan = this.getFlightPlan(forPlan);
-
-    const destinationMagVar = plan.destinationAirport
-      ? Facilities.getMagVar(plan.destinationAirport.location.lat, plan.destinationAirport.location.long)
-      : null;
-
-    this.updateTowerHeadwind(mag, dir, plan.destinationRunway?.magneticBearing ?? null);
-
-    const theta = MagVar.magneticToTrue(dir, destinationMagVar ?? 0) * MathUtils.DEGREES_TO_RADIANS;
-
-    const groundWind = { altitude: 0, vector: Vec2Math.setFromPolar(mag, theta, Vec2Math.create()), flags: 0 };
-    this.flightPlanService.setDescentWindEntry(0, groundWind, forPlan, false);
-
-    return true;
+    const success = await this.flightPlanService.setApproachWind(dir, mag, forPlan);
+    if (success && forPlan === FlightPlanIndex.Active) {
+      this.updateTowerHeadwind(mag, dir, plan.destinationRunway?.magneticBearing ?? null);
+    }
+    return success;
   }
 
   public setPerfApprTransAlt(s: string, forPlan: FlightPlanIndex): boolean {
@@ -4447,7 +4438,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
   //TODO: can this be an util? no
   public getMaxFlCorrected(): number | null {
     const maxFl = this.getMaxFL();
-    return maxFl !== null ? Math.min(maxFl, this.maximumRecommendedCruiseFlightLevel) : null;
+    return maxFl !== null ? Math.min(maxFl, FMCMainDisplay.maximumRecommendedCruiseFlightLevel) : null;
   }
 
   // only used by trySetMinDestFob
@@ -5131,8 +5122,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
         );
         const alternateWind = plan.performanceData.alternateWind.get();
 
-        const windComponent =
-          alternateWind !== null ? WindUtils.computeTailwindComponent(alternateWind, trueCourseToAlternate) : 0;
+        const windComponent = WindUtils.computeTailwindComponent(alternateWind, trueCourseToAlternate);
 
         const airDistance = A32NX_FuelPred.computeAirDistance(Math.round(distanceToAlt), windComponent);
 
@@ -5420,27 +5410,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
 
   computeAlternateCruiseLevel(forPlan: FlightPlanIndex): number | undefined {
     const plan = this.getFlightPlan(forPlan);
-    if (!plan) {
-      return undefined;
-    }
-
-    if (!plan.destinationAirport || !plan.alternateDestinationAirport) {
-      return undefined;
-    }
-
-    // TODO use actual flight plan distance rather than great circle distance
-    const distance = Avionics.Utils.computeGreatCircleDistance(
-      plan.destinationAirport.location,
-      plan.alternateDestinationAirport.location,
-    );
-
-    if (distance > 200) {
-      return 310;
-    } else if (distance > 100) {
-      return 220;
-    }
-
-    return 100;
+    return plan?.getAlternateCruiseLevel();
   }
 
   public goToFuelPredPage(forPlan: FlightPlanIndex) {
@@ -5476,7 +5446,13 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     if (!plan.pendingWindUplink.isWindUplinkReadyToInsert() && !plan.pendingWindUplink.isWindUplinkInProgress()) {
       plan.pendingWindUplink.onUplinkRequested();
 
-      const request = this.formatWindRequest(forPlan);
+      const request = formatWindRequest(
+        forPlan,
+        plan,
+        this.flightPhaseManager.phase,
+        this.guidanceController,
+        this.dataManager,
+      );
 
       try {
         const [status, uplink] = await this.atsu.receiveWindUplink(request, sentCallback);
@@ -5497,7 +5473,8 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
           uplink,
           plan,
           this.flightPhaseManager.phase,
-          FpmConfigs.A320_HONEYWELL_H3,
+          FpmConfigs.A320_HONEYWELL_H4,
+          FMCMainDisplay.maximumRecommendedCruiseFlightLevel,
         );
 
         this.addMessageToQueue(
@@ -5537,94 +5514,8 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     await this.flightPlanService.insertWindUplink(forPlan);
   }
 
-  private formatWindRequest(forPlan: FlightPlanIndex): WindRequestMessage {
-    const plan = this.getFlightPlan(forPlan);
-    if (!plan) {
-      return {};
-    }
-
-    const cruiseLevel = plan.performanceData.cruiseFlightLevel.get();
-    const phase = this.flightPhaseManager.phase;
-
-    const shouldRequestClimbWinds =
-      !plan.isActiveOrCopiedFromActive() ||
-      (phase >= FmgcFlightPhase.Preflight && phase <= FmgcFlightPhase.Takeoff) ||
-      phase === FmgcFlightPhase.Done;
-    const shouldRequestCruiseWinds =
-      !plan.isActiveOrCopiedFromActive() ||
-      (phase >= FmgcFlightPhase.Preflight && phase <= FmgcFlightPhase.Cruise) ||
-      phase === FmgcFlightPhase.Done;
-    const shouldRequestDescentWinds = plan.destinationAirport !== undefined;
-
-    const finalCruiseLevel = plan.allLegs.reduce(
-      (acc, leg) => (isLeg(leg) && leg.cruiseStep !== undefined ? Math.round(leg.cruiseStep.toAltitude / 100) : acc),
-      cruiseLevel,
-    );
-
-    const legPredictions =
-      forPlan === FlightPlanIndex.Active
-        ? this.guidanceController?.vnavDriver.mcduProfile?.waypointPredictions
-        : undefined;
-    const cruiseLegs = plan.allLegs.filter((leg, i) => {
-      if (!isLeg(leg) || !leg.isXF()) {
-        return false;
-      }
-
-      const legPrediction = legPredictions?.get(i);
-      return legPrediction !== undefined
-        ? legPrediction.profilePhase === ProfilePhase.Cruise
-        : leg.segment.class === SegmentClass.Enroute;
-    }) as FlightPlanLeg[];
-
-    let cruiseWinds: CruiseWindRequest | undefined = undefined;
-    if (shouldRequestCruiseWinds && cruiseLegs.length > 0) {
-      const propagatedWinds = this.flightPlanService.propagateWindsAt(0, [], forPlan);
-      const flightLevels = propagatedWinds.map((wind) => Math.round(wind.altitude / 100));
-
-      if (flightLevels.length === 0) {
-        if (cruiseLevel !== null) {
-          flightLevels.push(cruiseLevel);
-        }
-
-        plan.allLegs.forEach((leg) => {
-          if (isLeg(leg) && leg.cruiseStep !== undefined) {
-            const cruiseStep = Math.round(leg.cruiseStep.toAltitude / 100);
-
-            if (flightLevels.length < 4 && cruiseStep !== cruiseLevel && !flightLevels.includes(cruiseStep)) {
-              flightLevels.push(cruiseStep);
-            }
-          }
-        });
-      }
-
-      cruiseWinds = {
-        flightLevels,
-        waypoints: cruiseLegs.map((leg) => {
-          const isStoredWaypoint = this.dataManager.getStoredWaypointsByIdent(leg.ident).length > 0;
-
-          return isStoredWaypoint ? leg.definition.waypoint.location : leg.ident;
-        }),
-      };
-    }
-
-    let alternateWind = undefined;
-    if (plan.destinationAirport !== undefined && plan.alternateDestinationAirport !== undefined) {
-      alternateWind = {
-        destinationIcao: plan.destinationAirport.ident,
-        alternateIcao: plan.alternateDestinationAirport.ident,
-      };
-    }
-
-    return {
-      climbWindLevel: shouldRequestClimbWinds ? cruiseLevel : undefined,
-      cruiseWinds,
-      descentWindLevel: shouldRequestDescentWinds ? finalCruiseLevel ?? null : undefined,
-      alternateWind,
-    };
-  }
-
-  public getHistoryWinds(cruiseLevel: number | null) {
-    return this.historyWinds?.getRecordedWinds(cruiseLevel);
+  public async getHistoryWinds() {
+    return await this.flightPlanService.getHistoryWindsEntries(true);
   }
   // ---------------------------
   // CDUMainDisplay Types